Mock executable-find
When the tests run on Gitub actions, they are run in a container that has no
language servers installed. When the closure returned by `eglot-alternatives` is
invoked it will signal an error if no language server is found.

This fixes this problem by mocking `executable-find` so that it always returns
a non-nil value and `eglot--alternatives` never signals an exception